[Uan] Fix bug HPbar, Spawn player after enter combat,, independent modify player

+++ Lost relics/Assets/Scripts/UI/CharacterBar.cs
@@ -21,7 +21,7 @@ public class CharacterBar : MonoBehaviour
         
     }
 
-    public void updateHPBar(int maxHP, int currentHP)
+    public void updateHPBar(float maxHP, float currentHP)
     {
         slider.value = currentHP / maxHP;
     }
